#33 불필요한 패널 삭제
